Migrating main.cpp into Parse class
