+++ Library Search Engine/Library Search Engine/Film.hpp
@@ -87,6 +87,8 @@ public:
     
     std::string GetNotes() const;
     
+    friend class Parse;
+    
 };
