+++ Library Search Engine/Library Search Engine/Periodical.hpp
@@ -98,6 +98,8 @@ public:
     
     std::string GetNotes() const;
     
+    friend class Parse;
+    
 };
 
 #endif
