+++ Library Search Engine/Library Search Engine/Video.hpp
@@ -92,6 +92,8 @@ public:
 
     std::string GetNotes() const;
     
+    friend class Parse;
+    
 };
 
 #endif
